+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - #109 - Improve handling of Zayo notifications.
 - #110 - Improve handling of Telstra notifications.
+- #111 - Improve handling of EXA (GTT) notifications.
 
 ## v2.0.4 - 2021-11-04